backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
drf-yasg==1.21.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
joblib==1.5.3
kombu==5.6.2
lxml==6.0.2
nltk==3.9.2
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
regex==2026.1.15
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
textstat==0.7.12
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
xlsxwriter==3.2.9
